+++ packages/mini_program_sdk/lib/mini_program_page.dart
@@ -29,6 +29,7 @@ class MiniProgramPage extends StatelessWidget {
       hostBridge: resolvedRuntime.hostBridge,
       capabilityRegistry: resolvedRuntime.capabilityRegistry,
       backendConnector: resolvedRuntime.backendConnector,
+      authController: resolvedRuntime.authController,
       featureFlagEvaluator: resolvedRuntime.featureFlagEvaluator,
       assetCache: resolvedRuntime.cacheBundle.assetCache,
       manifestCache: resolvedRuntime.cacheBundle.manifestCache,